Fix no sudo use when installing multilib
When refactoring ansible script to avoid installing multilib packages
when executing on a aarch64 env a regression was introduced which breaks
installing such a lib on non-aarch64 environments. The regression
consists in not running the install command with sudo.

This commit fixes it by setting `become` as 'yes' so sudo is used when
installing that package.